MBX-3390 Update unit tests
