+++ MindboxTests/InApp/Tests/InAppConfigResponseTests/ConfigJsonStubs/ConfigTargetings/15-Targeting.json
@@ -1,0 +1,79 @@
+{
+    "settings": {
+      "operations": {
+        "viewProduct": {
+          "systemName": "viewProduct"
+        }
+      }
+    },
+    "inapps": [
+      {
+        "id": "1",
+        "sdkVersion": {
+          "min": 8,
+          "max": null
+        },
+        "targeting": {
+          "nodes": [
+            {
+              "internalId": "93",
+              "systemName": "test",
+              "$type": "apiMethodCall"
+            },
+            {
+                "kind": "negative",
+                "segmentationInternalId": "42",
+                "segmentationExternalId": "79ff415d-5bcf-45c8-ba78-bbf559083b5b",
+                "segmentExternalId": "79ff415d-5bcf-45c8-ba78-bbf559083b5b",
+                "$type": "segment"
+              }
+          ],
+          "$type": "and"
+        },
+        "form": {
+          "variants": [
+            {
+              "content": {
+                "background": {
+                  "layers": [
+                    {
+                      "action": {
+                        "intentPayload": "",
+                        "value": "",
+                        "$type": "redirectUrl"
+                      },
+                      "source": {
+                        "value": "https://mobpush-images.mindbox.ru/Mpush-test/456/5c7559bc-4095-4a66-9c43-8b8e3ec1f934.png",
+                        "$type": "url"
+                      },
+                      "$type": "image"
+                    }
+                  ]
+                },
+                "position": {
+                  "margin": {
+                    "kind": "dp",
+                    "top": 0.0,
+                    "right": 20.0,
+                    "left": 20.0,
+                    "bottom": 0.0
+                  },
+                  "gravity": {
+                    "horizontal": "center",
+                    "vertical": "top"
+                  }
+                },
+                "elements": [
+
+                ]
+              },
+              "imageUrl": "",
+              "redirectUrl": "",
+              "intentPayload": "",
+              "$type": "snackbar"
+            }
+          ]
+        }
+      }
+    ]
+  }
